requirements.txt
aiohttp>=2.2.5
ruamel.yaml
markdown2
more_itertools
alphabet_detector
beautifulsoup4
bleach
cachetools
CurrencyConverter
emoji
feedparser
future
geopy
gpytranslate
humanize
jikanpy
lxml
nltk
Pillow>=4.2.0
pretty_errors
psutil
psycopg2-binary
pynewtonmath
pyrate-limiter
python-telegram-bot==13.11
regex
requests
speedtest-cli
spongemock
sqlalchemy==1.3.20
pyrogram
ujson
wikipedia
zalgo-text
cloudscraper
